This branch modifies the existing solution to add 3 new projects: - FluentNHibernate2 - FluentNHibernate2.BackwardCompatibility - FluentNHibernate2.Testing
FluentNHibernate2 is the new core assembly. It contains the current progress on developing a semantic model for configuring nhibernate mappings. I also implemented a very small subset of the existing fluent api - this resides in FluentNHibernate2.BackwardCompatibility. The intention is to demonstrate that this rewrite supports the possibility of multiple fluent apis that each map to the same underlying model. To see the backward compatible fluent api in action, take a look at BackwardCompatibility/MusicMappingTester.cs. 

The intention of this branch is to act as an example of how the fluent aspects of this project can be seperated from the aspect of building nhibernate hbm xml. This branch has a LONG way to go before it could be considered as a possible replacement for the trunk. There are some significant issues to be worked out:

- Dealing with changes to the generated Hbm* classes is going to be hell. There are a few different ways this issue can be approached, but I have no idea if any of them have a chance of being successful. Need to look at the possibility of building the hbm graph in one hit, rather than incrementally, probably via the visitor pattern. It is also worth looking at the HbmGenerator project within NHibernate, see if changes can be made there, or, possibly to the partial classes with NHibernate.dll that augment the Hbm* classes.

- The trunk API supports enough of nhibernate mapping to be useful to some users. The api in this branch needs to be filled out significantly before the same can be said of it.

- No support for conventions or automapping. These are the most attractive features to many of the higher profile community members. This branch needs a proper treatment of these concepts before it can be taken seriously.

Final Note:
The projects include the "2" in the name simply because I found it most useful to maintain a solution that contains both the old code and the new code. The "2" will be removed should this branch progress far enough that is no longer necessary to include the original projects in the solution. So far I have only brought over functionality from the original projects when it was necessary, for example FluentNHibernate2/Cfg and FluentNHibernate2/Reflection. In both of these cases I tried to move the entire block of functionality over as is.


git-svn-id: https://fluent-nhibernate.googlecode.com/svn/branches/pb-rewrite@143 48f0ce17-cc52-0410-af8c-857c09b6549b