Fix for not-null not working.
This was because it was being set on the <property /> element, not the <column />. It used to work because we used to set the column="" attribute on the property element, but that was causing issue with custom sql-type's. Without the column attribute, the not-null attribute is ignored unless it is on the column element.

From: James Gregory <james@jagregory.com>

git-svn-id: https://fluent-nhibernate.googlecode.com/svn/trunk@330 48f0ce17-cc52-0410-af8c-857c09b6549b