Resolves issue 131 - Added regression test
Changed CachePart on the ClassMap to be stored exclusively in the Parts collection. This allowed removal of the AddPart(Cache) from the CreateMapping() method call, which was the culprit that was adding the CachePart multiple times during a single configuration sequence.

Added a global Cache.Provider_Class config option to the test project which allows a session to be built during the config tests when a ClassMap has been defined that contains a CachePart (such as CachedRecordMap is currently set up to do).

git-svn-id: https://fluent-nhibernate.googlecode.com/svn/trunk@452 48f0ce17-cc52-0410-af8c-857c09b6549b