Implemented support for defining subclasses outside of their parent.
It is now possible to define a subclass in the same way that you do a normal class, you just derive from SubclassMap<YourEntity> in the same manner as you would ClassMap<YourEntity>.

One important thing to know is that there isn't a distinction between subclasses and joined-subclasses, you map them both the same way. It will be determined by inspecting the parent mapping as to whether the subclass you've defined is joined or not; this is done by looking for a discriminator definition on the parent. To use subclasses instead of joined-subclasses just call DiscriminateSubclassesOnColumn in your parent's ClassMap.