Introduced new way to configure the automapper.
The automapper can (and should) now be configured by supplying a
configuration instance -- an implementation of IAutomappingConfiguration
-- which specifies how the automapper should behave when mapping the
domain.

All configuration options from the Setup action have been migrated,
along with the Where method.

The Setup-based usage is still available (although depreciated), but you
can't mix-and-match the types; if a configuration instance is supplied
the Setup and Where methods will throw.

The automapper has also been modified to act on all members of a class,
ra than just properties. The default is still to use properties, but the
ShouldMap(Member) method can be overridden to use fields or any
combination.