Implemented OrderByOnRelationshipElement method on ManyToManyPart for setting the order-by attribute on a many-to-many element, instead of on the collection element (e.g. bag, set, etc).
Apparently the order-by attribute has different semantics depending on which element its specified on. According to David T:
"If I understand it correctly, when on the bag, the order by will work on the columns of the association table, and when on the many-to-many, it will work on the referenced table's columns."

Hoping a user will suggest a better name for this method.