Fix for automapping inheritance order.
The change made in 894d218 which added an order to the type sources had
the side-effect of breaking the implicit parent-first order of types
that the automapper relied on. This fix introduces an order on the
"depth" of a type, aka the number of steps required to get back to
object from the current type. Types with the least depth are mapped
first, meaning parents are mapped before their children.