Support m2o-o2m self-referential relationships
The automapper now supports bi-directional many-to-one one-to-many
self-referential relationships. These are typically the tree structure
style classes, where class A has a Parent property of A, and a Children
collection of A.

To properly implement this fix, the work for #142 had to be done too.
This was to improve the way column names are generated for
bi-directional relationships, by making sure both sides share the same
key columns.

Fixes #115 and #142