Fixed multiple instance of same ComponentMap bug
A setup satisfying the following:

  * An entity with two components of the same type, each with it's own
    prefix
  * A ComponentMap for the above component type, with explicit columns
    being used

Would cause columns to be mapped with both prefixes. This was due to
effectively the same instance of component being assigned to each
property. Without the explicit columns, default ones are used which
aren't shared between the instances; however, with the prefixes both
components shared a column instance. When the prefix visitor was applied
to the shared column, it received both prefxies.

I've implemented deep cloning using binary serialization to remedy this.
I was against using this method previously because I think it's hacky,
but it seems like the only acceptable solution besides hand-rolling a
deep copy mechanism (which I'm not doing).

Fixes #173