Fixed issue whereby an inherited autoproperty with private setter was not correctly mapped with access=backfield.
This was due to a bug in the logic for determining whether a member is an autoproperty. It incorrectly assumed that an auto property had to have an accessible set method, which is not the case if the autoproperty was declared on a base type with a private setter, but discovered when reflecting on a derived type.