Can ignore intermediary subclasses in automappings
You can now ignore classes that are in the centre of an inheritance
hierarchy using the automapper. Previously if you did this, the
automapper couldn't find the link between the child and parent (because
the child's immediate parent wasn't being mapped).

Due to this change, you can just make the intermediary class abstract
and it will be ignored automatically and the hierarchy will continue to
function; otherwise, you can use the explicit IgnoreBase<T> method to
ignore non-abstract classes.

Fixes #128