Added no-proxy lazy support
Updated the References and HasOne methods to include support for
no-proxy lazy loading. This is done through an overload with a Laziness
parameter, to avoid breaking changes.

LazyLoad(Laziness.NoProxy);

The Laziness class can also be instantiated by the user with their own
value; this will be handy for if NHibernate adds any more values to this
enumerable before we can catch up.

Fixes #183