Fix for weird key column issue
1.1 introduced a regression where entities that are the target of
multiple Reference mappings from the same source would end up with
multiple key columns being defined based on the property names.

The source of this was the RelationshipKeyPairingVisitor, which frankly
I have no idea what it was trying to do. I think it might have something
to do with composite keys and relationships, but the usage isn't very
clear. For the time being I've removed this functionality.