Update src/FluentNHibernate/MappingModel/Collections/LayeredColumns.cs
This class uses a Hashset intern. This is no problem with .Net, but mono (at least version 2.10.8 and sooner) has a incorrect Hashset implementation, so this class is not workin correct with the mono runtime. As long the hashset in mono is not fixed, I suggest to use a Dictionary here.

Using a Dictionary in this class insteas